Solved several bugs, but the first layer features are still buggy

+++ Libs/cl_kernels/conv_w_update.cl
@@ -50,8 +50,8 @@ __kernel void conv_w_update(__global int *ts, __global int *surf_x_b,
                                 + loc_idx;
                             
                 weights[lin_idx] = weights[lin_idx] + 
-                                   (0.01f)*S[i_file]*lrate*dweights[lin_idx] + 
-                                   dS[i_file]*lrate*dweights[lin_idx];
+                                   ((float)(0.01f)*S[i_file]*lrate)*dweights[lin_idx] + 
+                                   ((float)dS[i_file]*lrate)*dweights[lin_idx];
                 
                 }  
             }
